refactor: restructure history page folder, split server/client code, update styles
